Separate FloatOps.Generated from IntOps.Generated
